requirements.txt
asttokens==2.0.5
async-generator==1.10
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
colorama==0.4.5
cryptography==37.0.4
cycler==0.11.0
debugpy==1.6.2
decorator==5.1.1
entrypoints==0.4
executing==0.8.3
fonttools==4.34.4
h11==0.13.0
idna==3.3
ipykernel==6.15.1
ipython==8.4.0
jedi==0.18.1
jupyter-client==7.3.4
jupyter-core==4.11.1
kiwisolver==1.4.4
lxml==4.9.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
nest-asyncio==1.5.5
numpy==1.23.1
outcome==1.2.0
packaging==21.3
pandas==1.4.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.30
psutil==5.9.1
pure-eval==0.2.2
pycparser==2.21
Pygments==2.12.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
pywin32==304
pyzmq==23.2.0
requests==2.28.1
selenium==4.3.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
stack-data==0.3.0
tornado==6.2
tqdm==4.64.0
traitlets==5.3.0
trio==0.21.0
trio-websocket==0.9.2
urllib3==1.26.10
wcwidth==0.2.5
wordcloud==1.8.2.2
wsproto==1.1.0
